[refactor] AppMasterNavigationLayout: promote it to replace FAZNCodeChallengeApp as it does nothing than passing values
